[refactor] WBJ-566 Dashboard page components refactored
