+++ src/pages/docs/hooks/index.ts
@@ -1,1 +1,2 @@
 export { default as useModal } from './useModal';
+export { default as useMoment } from './useMoment';
